Fix failures of locale_match!
This change adds postprocessing of parsed constants to determine if
values are present in all locales (optional), or ensures that the types
are always the same. For example, if a value was previously registerd as
`i64` for one locale _and_ a `String` for another, it will now always be
a `Sttring, it will now always be a `String`.

Similarly, when certain values are present in some locales but not others,
the types will be wrapped in `Option`.

When a whole module is missing, a dummy one will be created with all
`const`s set to `None`.

All of this should fix the compilation errors when trying to use
`locale_match!` with locales that previously had diverging types.

This is most likely _not_ a backwards compatible change, still I'm going
to double check that nothing breaks in `chrono`.

Closes #4.